requirements.txt
asgiref==3.2.10
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Django==2.2.16
django-appconf==1.0.4
django-imagekit==4.0.2
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-vision==2.2.0
googleapis-common-protos==1.52.0
grpcio==1.35.0
idna==2.10
libcst==0.3.17
mypy-extensions==0.4.3
packaging==20.9
pilkit==2.0
Pillow==8.1.0
proto-plus==1.13.0
protobuf==3.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytz==2020.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.1
six==1.15.0
sqlparse==0.3.1
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.26.3
